Edited constructor and user input
Wrote part of the menu with the user input into the constructor, however it's incomplete and not working right now.

+++ Binckes Bakfiets Console App V2 Portfolio/ShopItems.cs
@@ -8,7 +8,7 @@ namespace Binckes_Bakfiets_Console_App_V2_Portfolio
 {
     public class ShopItems
     {
-        public int ItemPrice { get; set; }
+        public double ItemPrice { get; set; }
 
 
     }
